Старая версия протокола

+++ MessageSender.h
@@ -23,7 +23,7 @@ public:
     static uint16_t computeCRC(const uint8_t* data, size_t length);
 private:    
     void runImpl(std::stop_token stoken, HardwareManager* manager);
-    uint16_t controlSum() const;
+    uint8_t controlSum() const;
     struct Impl;
     std::unique_ptr<Impl> _impl;
 };
